src/old/requirements.txt
python-chess
chess
# cv2
numpy
# json
# chess.svg
pyserial
cairosvg
firebase_admin
flask
flask-cors